+++ src/@next/components/atoms/Checkbox/test.tsx
@@ -1,0 +1,15 @@
+import { shallow } from "enzyme";
+import "jest-styled-components";
+import React from "react";
+
+import { Checkbox } from ".";
+
+describe("<Checkbox />", () => {
+  it("exists", () => {
+    const wrapper = shallow(
+      <Checkbox name="default-checkbox">Checkbox with label</Checkbox>
+    );
+
+    expect(wrapper.exists()).toEqual(true);
+  });
+});