+++ src/@next/components/organisms/index.ts
@@ -10,3 +10,4 @@ export * from "./AddressFormModal";
 export * from "./FilterSidebar";
 export * from "./ProductVariantPicker";
 export * from "./SelectSidebar";
+export * from "./DiscountForm";