+++ src/@sdk/mutations/checkout.ts
@@ -95,3 +95,43 @@ export const updateCheckoutShippingAddressMutation = gql`
     }
   }
 `;
+
+export const addCheckoutPromoCode = gql`
+  ${checkoutFragment}
+  mutation AddCheckoutPromoCode($checkoutId: ID!, $promoCode: String!) {
+    checkoutAddPromoCode(checkoutId: $checkoutId, promoCode: $promoCode) {
+      checkout {
+        ...Checkout
+      }
+      errors {
+        field
+        message
+      }
+      checkoutErrors {
+        field
+        message
+        code
+      }
+    }
+  }
+`;
+
+export const removeCheckoutPromoCode = gql`
+  ${checkoutFragment}
+  mutation RemoveCheckoutPromoCode($checkoutId: ID!, $promoCode: String!) {
+    checkoutRemovePromoCode(checkoutId: $checkoutId, promoCode: $promoCode) {
+      checkout {
+        ...Checkout
+      }
+      errors {
+        field
+        message
+      }
+      checkoutErrors {
+        field
+        message
+        code
+      }
+    }
+  }
+`;